// PP-CoE-CodeApp/src/generated/models/PowerPlatformforAdminsV2Model.ts
export interface EnvironmentGroup {
  id?: string;
  displayName?: string;
  description?: string;
  parentGroupId?: string;
  childrenGroupIds?: string[];
  createdTime?: string;
  createdBy?: Principal;
  lastModifiedTime?: string;
  lastModifiedBy?: Principal;
}
// ...
export interface MgGovEnvironmentValues {
  // The ID of the environment.
  id?: string;
  // The type of the environment.
  type?: string;
  // The name of the environment.
  name?: string;
}

export interface MgGovErrorResponse {
  error?: {
    // Error code.
    code?: string;
    // Error message.
    message?: string;
  };
}

export type MgGovFilterType = string;

export interface MgGovODataResponse {
  // The array of Rule Sets.
  value?: RuleSetDto[];
  // The link to the next page of results.
  "@odata.nextLink"?: string;
}

export interface MgGovPolicyEnvironmentFilter {
  type?: MgGovFilterType;
  // The environment information.
  values?: MgGovEnvironmentValues[];
}

export type MgGovResourceType = string;

export interface MgGovRule {
  // The rule ID.
  id?: string;
  // The rule value.
  value?: string;
}

export type MgGovRuleSetType = string;
// ...
export interface Principal {
  id?: string;
  displayName?: string;
  email?: string;
  type?: string;
  tenantId?: string;
  userPrincipalName?: string;
}
// ...
export interface RuleSetDto {
  // The ID of the Rule Set.
  id?: string;
  // The last modified timestamp.
  lastModified?: string;
  environmentFilter?: MgGovPolicyEnvironmentFilter;
  // The Rule Set parameters.
  parameters?: RuleSetParameters[];
}

export interface RuleSetParameters {
  type: MgGovRuleSetType;
  resourceType: MgGovResourceType;
  // The rule values.
  value?: MgGovRule[];
}